Convert all movement modes and include bonuses
This solves an issue originally reported by DragoSinOfWrath on
Discord.

Some speeds listed in racial traits are written as "equal to your
X speed" without giving a flat value, so we needed new logic to
parse those values.

Additionally, items that give bonuses to speed or particular types
of speed were not factored into the conversion of movement modes.