feat: converter options
Added types and an extra "options" param to the convertCharacter function so that only the properties specified in the options object are converted.

This is going to aid in writing tests as now we don't need to run all code paths when testing a specific fragment of the conversion code.