👔 Adjust XP tracker to show `current total / next level`
Instead of adjusting the value relative to the current level (which
would look nice on the bar), this instead shows the current total
experience as the value, and the next level as the max. This is closer
to what D&D Beyond shows, although it has a nicer bar that shows the
progress to the next level. We don't really have a way to do that solely
with a tracker at the moment, though.